% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_postcodes}
\alias{clean_postcodes}
\title{Clean postcodes}
\usage{
clean_postcodes(x)
}
\arguments{
\item{x}{A \code{vector} or data column of \code{character} class.}
}
\description{
Standardize Brazilian postal codes (CEP) to 8-digit, all-numeric format.
}
