% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_rais_names}
\alias{clean_rais_names}
\title{Clean character vectors}
\usage{
clean_rais_names(x)
}
\arguments{
\item{x}{A \code{vector} or data column of \code{character} class.}
}
\description{
Removes any diacritics present on strings, extra spaces, leading zeroes, and converts them to lowercase.
}
\details{
Auxiliary function to \code{street_sweeper}. This function standardizes street names, which increases the chances of
matching firms addresses. Can also be used with any string function, e.g. worker name.
@export
}
