% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_repair}
\alias{date_repair}
\title{Auxiliary functions
Fix dates to include missing zeroes}
\usage{
date_repair(x)
}
\arguments{
\item{x}{A \code{character} vector.}
}
\description{
Auxiliary functions
Fix dates to include missing zeroes
}
\details{
The function will result in a \code{character} vector in the standard Brazilian "DDMMAAAA" format. It can be easily
converted into a functional date variable, e.g. with \code{lubridate::dmy(x)}.
}
