% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decimal_repair}
\alias{decimal_repair}
\title{Replace decimal mark for numbers read as characters}
\usage{
decimal_repair(x)
}
\arguments{
\item{x}{A \code{vector} of soon-to-be numbers.}
}
\description{
Replace decimal mark for numbers read as characters
}
