% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic_firms.R
\docType{data}
\name{dic_firms}
\alias{dic_firms}
\title{List of firm dataset variables}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 58 rows and 6 columns.
}
\usage{
dic_firms
}
\description{
For each variable, we list its coverage period, eventual years skipped, the original name (\code{alias}) and the
standard variable name that we adopted (\code{new_name}). Note that \code{new_name} repeats because data layout changes
throughout the years.
}
\author{
Arthur Alvarenga \url{baarthur.github.io}
}
\keyword{datasets}
