% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_rais_dataset.R
\name{open_rais_dataset}
\alias{open_rais_dataset}
\title{Read raw RAIS data as an arrow dataset}
\usage{
open_rais_dataset(
  file,
  year,
  worker_dataset = TRUE,
  filename,
  delim = NULL,
  ...
)
}
\arguments{
\item{file}{RAIS dataset in \code{.txt} format.}

\item{year}{Fiscal year of the file, necessary to determine column types, which
columns to read, and other time-specific tidying.}

\item{worker_dataset}{When reading establishment data, set \code{worker_dataset = FALSE}. Default
option is to read work relationship data (\code{worker_dataset = TRUE}).}

\item{filename}{if \code{NULL}, returns file; otherwise, provide a string for the destination file.
Must end with with ".parquet".}

\item{delim}{Delimiter type. If \code{NULL} (the default), semicolon (\verb{;}) will be used.}

\item{\dots}{Other variables passed on to \code{\link[arrow:open_delim_dataset]{arrow::open_delim_dataset()}}}
}
\description{
Read raw RAIS data as an arrow dataset
}
\details{
Replace a time- and memory-consuming pipeline with a leaner approach. A wrapper around
\code{\link[arrow:open_delim_dataset]{arrow::open_delim_dataset()}} to automate the first steps of processing RAIS data, such as
choosing the right delimiter and making sure that leading zeroes in identifier columns are not
removed.
}
