% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rais_to_parquet.R
\name{rais_to_parquet}
\alias{rais_to_parquet}
\title{Save raw RAIS data on parquet format}
\usage{
rais_to_parquet(
  file,
  year,
  columns = NULL,
  worker_dataset = TRUE,
  filename,
  delim = NULL,
  ...
)
}
\arguments{
\item{file}{RAIS dataset in \code{.txt} format.}

\item{year}{Fiscal year of the file, necessary to determine column types, which
columns to read, and other time-specific tidying.}

\item{columns}{A \code{vector} of variables to read.}

\item{worker_dataset}{When reading establishment data, set \code{worker_dataset = FALSE}. Default
option is to read work relationship data (\code{worker_dataset = TRUE}).}

\item{filename}{destination file. must end with ".parquet".}

\item{delim}{Delimiter type. If \code{NULL} (the default), semicolon (;) will be used.}

\item{\dots}{Other variables passed on to \code{readr::read_delim()}}
}
\description{
Save raw RAIS data on parquet format
}
