% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_firms.R
\name{read_firms}
\alias{read_firms}
\title{Read Establishments from RAIS dataset.}
\usage{
read_firms(
  file,
  firm_filter = NULL,
  muni_filter = NULL,
  cep_filter = NULL,
  address_filter = NULL,
  street_filter = NULL,
  delim = NULL,
  year = NULL,
  rais_negativa = TRUE,
  discard = T,
  discard_other = NULL,
  col_select = NULL,
  ...
)
}
\arguments{
\item{file}{RAIS dataset in \code{.csv} format.}

\item{firm_filter}{A \code{vector} of firm CNPJs to use as filter.}

\item{muni_filter}{A \code{vector} of municipalities in 6-digit IBGE code to use as filter.}

\item{cep_filter}{A \code{vector} of postal codes to use as filter.}

\item{address_filter}{A \code{vector} of addresses to use as filter for identified data up to 2010.}

\item{street_filter}{A \code{vector} of street names to use as filter for identified data from 2011 onwards.}

\item{delim}{Delimiter type. If \code{NULL} (the default), semicolon (;) will be used.}

\item{year}{Reserved for future use.}

\item{rais_negativa}{Should only establishments with active employees in that year  ("Rais Negativa") be selected?
Default is \code{TRUE}.}

\item{discard}{Discard a pre-selected set of variables? Default is \code{TRUE}.}

\item{discard_other}{A \code{vector} of variables to discard instead of (or in addition to) the default set.}

\item{col_select}{A \code{vector} of variables to be selected.}

\item{\dots}{Other variables passed on to readr::read_delim}
}
\value{
A \code{tibble}
}
\description{
Read Establishments from RAIS dataset.
}
