% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_workers.R
\name{read_workers}
\alias{read_workers}
\title{Read Workers from RAIS dataset.}
\usage{
read_workers(
  file,
  firm_filter = NULL,
  muni_filter = NULL,
  delim = NULL,
  year = NULL,
  vinculo_ativo = F,
  discard = T,
  discard_other = NULL,
  col_select = NULL,
  ...
)
}
\arguments{
\item{file}{RAIS dataset in \code{.csv} format.}

\item{firm_filter}{A \code{vector} of firm CNPJs to use as filter.}

\item{muni_filter}{A \code{vector} of municipalities in 6-digit IBGE code to use as filter.}

\item{delim}{Delimiter type. If \code{NULL} (the default), semicolon (;) will be used.}

\item{year}{Reserved for future use.}

\item{vinculo_ativo}{Should only workers that were active at the end of the year be selected?
Default is \code{TRUE}.}

\item{discard}{Discard a pre-selected set of variables? Default is \code{TRUE}.}

\item{discard_other}{A \code{vector} of variables to discard instead of (or in addition to) the default set.}

\item{col_select}{A \code{vector} of variables to be selected.}

\item{\dots}{Other variables passed on to readr::read_delim}
}
\value{
A \code{tibble}
}
\description{
Read Workers from RAIS dataset.
}
