% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrais-package.R
\docType{package}
\name{rrais-package}
\alias{rrais}
\alias{rrais-package}
\title{rrais: Tidy and Manipulate the Brazilian Employer-Employee Dataset}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Arthur Bazolli \email{baarthur0@outlook.com}

}
\keyword{internal}
