% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/street_sweeper.R
\name{street_sweeper}
\alias{street_sweeper}
\title{Tidy addresses from RAIS dataset}
\usage{
street_sweeper(df, addr, split = TRUE, keep_original = TRUE)
}
\arguments{
\item{df}{A \code{data.frame} or any object coercible to it.}

\item{addr}{\verb{<data-masking>} Column in \code{df} with addresses.}

\item{split}{If \code{TRUE} (default), separates the street type (e.g., "Rua", "Avenida"), name, number, and "other"
(floor, unit etc) into specific columns.}

\item{keep_original}{If \code{TRUE} (default), original \code{addr} data is preserved.}
}
\description{
\code{street_sweeper()} sweeps the dirt out of street names, removing misplaced characters and breaking data into
useful chunks.
}
\details{
\code{street_sweeper} uses a predetermined set of street types, including common typos (such as Ra instead of Rua), to
separate \code{street_type} from the full address.
This is useful for matching
addresses across firms or along time and for geocoding.
}
\examples{
df <- tibble::tibble(
  addr_full = c("'RUA DOS BOBOS, 687 APTO 1904 NESTA", "AV. GRAVIDA DE TAUBATE, 4159  TERREO",
                "RUA 1 DE ABRIL, 7  NESTA", "-RUA SANDUICHE-ICHE, 1830  TERREO",
                ",ALAMEDA VOVOZINHA, 1577  NESTA", "AV ALA ALAMBIQUE, 173",
                "RUA Z  QUADRA X LOTE +++9",
                "R.DOIS, 15 APTO 1", "AV.21 DE ABRIL 345", "R. MARIA, 16 APTO 15",
                "AVENIDA 21DE ABRIL, 42 BL A AP 1", "R.15, 123", "PRAÇA XV, S/N",
                "BR-356 SN", "PC TENENTE JOSE, 98")
)

street_sweeper(df, addr_full)

street_sweeper(df, addr_full, split = FALSE)

street_sweeper(df, addr_full, keep_original = FALSE)

street_sweeper(df, addr_full, keep_original = FALSE, split = FALSE)
}
