% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rais.R
\name{read_rais}
\alias{read_rais}
\title{Read Establishments and Work Relationships from RAIS dataset.}
\usage{
read_rais(
  file,
  year,
  worker_dataset = TRUE,
  columns = NULL,
  remove_rais_negativa = TRUE,
  vinculo_ativo = TRUE,
  muni_filter = NULL,
  cep_filter = NULL,
  address_filter = NULL,
  street_filter = NULL,
  firm_filter = NULL,
  delim = NULL,
  ...
)
}
\arguments{
\item{file}{RAIS dataset in \code{.csv} format.}

\item{year}{Reserved for future use.}

\item{worker_dataset}{When reading establishment data, set \code{worker_dataset = FALSE}. Default option is to read
work relationship data (\code{worker_dataset = TRUE}).}

\item{columns}{A \code{vector} of variables to be selected.}

\item{remove_rais_negativa}{Should only establishments with active employees in that year  ("Rais Negativa") be selected?
Default is \code{TRUE}.}

\item{vinculo_ativo}{Should only workers that were active at the end of the year be selected?
Default is \code{TRUE}.}

\item{muni_filter}{A \code{vector} of municipalities in 6-digit IBGE code to use as filter.}

\item{cep_filter}{A \code{vector} of postal codes to use as filter.}

\item{address_filter}{A \code{vector} of addresses to use as filter for identified data up to 2010.}

\item{street_filter}{A \code{vector} of street names to use as filter for identified data from 2011 onwards.}

\item{firm_filter}{A \code{vector} of firm CNPJs to use as filter, if using the identified dataset.}

\item{delim}{Delimiter type. If \code{NULL} (the default), semicolon (;) will be used.}

\item{\dots}{Other variables passed on to \code{readr::read_delim()}}
}
\value{
A \code{data.table}
}
\description{
\code{read_rais()} reads raw RAIS data and generates consistent variables for different years.
}
\details{
\code{read_rais()} first uses \code{readr::read_delim()} to read the .txt RAIS files. Then, it standardizes variables
across years with unified names and column types and repairs continuous values and dates, making data robust to
the various layout changes between 1985 and 2020.

This standardization enhances data analysis, e.g. by enabling the user to stack different years in panel data
format. The multiple available filters reduces memory use, by getting rid of unnecessary data before tidying.

Since 2011, RAIS split the address variable into three columns (street name, street number, and neighborhood).
We opted not to unify these columns, since it would imply a moderate level of arbitrariness into how we split
pre-2011 data or unify post-2011 data. If needed, use the companion \code{street_sweeper()} function to clean street
names following a consistent pattern, useful for geolocating or tracking firm address changes.
}
\examples{
# read_rais example
\dontrun{
  library(rrais)

  # simplest use: provide file and year

  path <- "path/to/rais_firms.txt"

  df <- read_rais(file = path, year = 2003, worker_dataset = FALSE)


  # filter for a vector of cities in IBGE-6 format
  cities <- c(355030, 310620, 320530)
  df <- read_rais(path, 2013, muni_filter = cities, worker_dataset = FALSE)


  # select a set of columns
  ## data("dic_firms") # optional: lookup variable names
  cols <- c("cbo_94", "cbo_2002", "genero", "rem_med", "vinculo_ativo_31_12")

  df <- read_rais("path/to/rais_worker.txt", 2015, columns = cols, n_max = 10)

}
}
\seealso{
\code{\link[readr:read_delim]{readr::read_delim()}} for the arguments passed on to it and the
\href{https://www.gov.br/trabalho-e-emprego/pt-br}{Ministry of Labor and Employment (MTE)} website for
information on RAIS.
}
