% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/street_sweeper.R
\name{street_sweeper}
\alias{street_sweeper}
\title{Tidy addresses from RAIS dataset}
\usage{
street_sweeper(df, addr, split = TRUE, keep_original = TRUE)
}
\arguments{
\item{df}{A \code{data.frame} or any object coercible to it.}

\item{addr}{\verb{<data-masking>} Column in \code{df} with addresses.}

\item{keep_original}{If \code{TRUE} (default), original \code{addr} data is preserved.}

\item{sep_type}{If \code{TRUE} (default), separates the street type (e.g., "Rua", "Avenida") into a specific column.}

\item{sep_number}{If \code{TRUE} (default), separates the address number into a specific column.}
}
\description{
Tidy addresses from RAIS dataset
}
\details{
\code{street_sweeper} sweeps the dirt out of street names, by fixing common typos such as "Ra" instead of "Rua"
and standardizes street types (e.g, "Avenida", "Ave." "Av" all become the same). This is useful for matching
addresses across firms or along time and for geocoding.
@example inst/examples/street_sweeper.R
}
